Finish Work on Setup Redress
The setup script isn't finished, but the major issues have been righted.

This commit adds some verbosity to the UI feedback (print statements),
and the setup script also now checks for the existance of the
config['default_view'] file.

Decided to ignore the following bullet point in the redress issue:
* Have setup script ensure that the document root for Apache is correct.
Because apachectl does this already.