Added $_SERVER['PROJECT_DIR'] to chameleon Script; Broke off Header and Apache Config
The Global accomadates the moving of the chameleon script into the bin
directory. Now the script knows the absolute path to many of the files
and directories referred to in the script -- not yet all I think.

The chameleon script is large, so I broke it off into two more files.